remove mutating webhook from Seldon Core Operator
