+++ helm-charts/seldon-core-operator/templates/clusterrole_seldon-webhook-role.yaml
@@ -14,7 +14,6 @@ rules:
 - apiGroups:
   - admissionregistration.k8s.io
   resources:
-  - mutatingwebhookconfigurations
   - validatingwebhookconfigurations
   verbs:
   - get
