+++ helm-charts/seldon-core-operator/templates/webhook.yaml
@@ -184,173 +184,5 @@ webhooks:
     resources:
     - seldondeployments
   sideEffects: None
----
-
-apiVersion: admissionregistration.k8s.io/v1beta1
-kind: MutatingWebhookConfiguration
-metadata:
-  annotations:
-    cert-manager.io/inject-ca-from: '{{ include "seldon.namespace" . }}/seldon-serving-cert'
-  creationTimestamp: null
-  labels:
-    app: seldon
-    app.kubernetes.io/instance: '{{ .Release.Name }}'
-    app.kubernetes.io/name: '{{ include "seldon.name" . }}'
-    app.kubernetes.io/version: '{{ .Chart.Version }}'
-  name: seldon-mutating-webhook-configuration-{{ include "seldon.namespace" . }}
-webhooks:
-- clientConfig:
-    caBundle: '{{ $ca.Cert | b64enc }}'
-    service:
-      name: seldon-webhook-service
-      namespace: '{{ include "seldon.namespace" . }}'
-      path: /mutate-machinelearning-seldon-io-v1-seldondeployment
-  failurePolicy: Fail
-  name: v1.mseldondeployment.kb.io
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.singleNamespace }}
-  namespaceSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- if .Values.kubeflow }}
-    matchLabels:
-      serving.kubeflow.org/inferenceservice: enabled
-{{- end }}
-{{- end }}
-{{- end }}
-{{- if .Values.singleNamespace }}
-  namespaceSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ include "seldon.namespace" . }}
-{{- end }}
-{{- if not .Values.kubeflow }}
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.controllerId }}
-  objectSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- end }}
-{{- end }}
-{{- if .Values.controllerId }}
-  objectSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ .Values.controllerId }}
-{{- end }}
-{{- end }}
-  rules:
-  - apiGroups:
-    - machinelearning.seldon.io
-    apiVersions:
-    - v1
-    operations:
-    - CREATE
-    - UPDATE
-    resources:
-    - seldondeployments
-  sideEffects: None
-- clientConfig:
-    caBundle: '{{ $ca.Cert | b64enc }}'
-    service:
-      name: seldon-webhook-service
-      namespace: '{{ include "seldon.namespace" . }}'
-      path: /mutate-machinelearning-seldon-io-v1alpha2-seldondeployment
-  failurePolicy: Fail
-  name: v1alpha2.mseldondeployment.kb.io
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.singleNamespace }}
-  namespaceSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- if .Values.kubeflow }}
-    matchLabels:
-      serving.kubeflow.org/inferenceservice: enabled
-{{- end }}
-{{- end }}
-{{- end }}
-{{- if .Values.singleNamespace }}
-  namespaceSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ include "seldon.namespace" . }}
-{{- end }}
-{{- if not .Values.kubeflow }}
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.controllerId }}
-  objectSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- end }}
-{{- end }}
-{{- if .Values.controllerId }}
-  objectSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ .Values.controllerId }}
-{{- end }}
-{{- end }}
-  rules:
-  - apiGroups:
-    - machinelearning.seldon.io
-    apiVersions:
-    - v1alpha2
-    operations:
-    - CREATE
-    - UPDATE
-    resources:
-    - seldondeployments
-  sideEffects: None
-- clientConfig:
-    caBundle: '{{ $ca.Cert | b64enc }}'
-    service:
-      name: seldon-webhook-service
-      namespace: '{{ include "seldon.namespace" . }}'
-      path: /mutate-machinelearning-seldon-io-v1alpha3-seldondeployment
-  failurePolicy: Fail
-  name: v1alpha3.mseldondeployment.kb.io
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.singleNamespace }}
-  namespaceSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- if .Values.kubeflow }}
-    matchLabels:
-      serving.kubeflow.org/inferenceservice: enabled
-{{- end }}
-{{- end }}
-{{- end }}
-{{- if .Values.singleNamespace }}
-  namespaceSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ include "seldon.namespace" . }}
-{{- end }}
-{{- if not .Values.kubeflow }}
-{{- if semverCompare ">=1.15.0" .Capabilities.KubeVersion.GitVersion }}
-{{- if not .Values.controllerId }}
-  objectSelector:
-    matchExpressions:
-    - key: seldon.io/controller-id
-      operator: DoesNotExist
-{{- end }}
-{{- end }}
-{{- if .Values.controllerId }}
-  objectSelector:
-    matchLabels:
-      seldon.io/controller-id: {{ .Values.controllerId }}
-{{- end }}
-{{- end }}
-  rules:
-  - apiGroups:
-    - machinelearning.seldon.io
-    apiVersions:
-    - v1alpha3
-    operations:
-    - CREATE
-    - UPDATE
-    resources:
-    - seldondeployments
-  sideEffects: None
 
 {{- end }}
