+++ operator/apis/machinelearning.seldon.io/v1alpha2/seldondeployment_webhook.go
@@ -37,20 +37,6 @@ func (r *SeldonDeployment) SetupWebhookWithManager(mgr ctrl.Manager) error {
 		Complete()
 }
 
-var _ webhook.Defaulter = &SeldonDeployment{}
-
-// +kubebuilder:webhook:path=/mutate-machinelearning-seldon-io-v1alpha2-seldondeployment,mutating=true,failurePolicy=fail,sideEffects=None,groups=machinelearning.seldon.io,resources=seldondeployments,verbs=create;update,versions=v1alpha2,name=v1alpha2.mseldondeployment.kb.io
-
-// Default implements webhook.Defaulter so a webhook will be registered for the type
-func (r *SeldonDeployment) Default() {
-	seldondeploymentlog.Info("Defaulting v1alpha2 webhook called", "name", r.Name)
-
-	if r.ObjectMeta.Namespace == "" {
-		r.ObjectMeta.Namespace = "default"
-	}
-	r.Spec.DefaultSeldonDeployment(r.Name, r.ObjectMeta.Namespace)
-}
-
 // TODO(user): change verbs to "verbs=create;update;delete" if you want to enable deletion validation.
 // +kubebuilder:webhook:verbs=create;update,path=/validate-machinelearning-seldon-io-v1alpha2-seldondeployment,mutating=false,failurePolicy=fail,sideEffects=None,groups=machinelearning.seldon.io,resources=seldondeployments,versions=v1alpha2,name=v1alpha2.vseldondeployment.kb.io
 
@@ -59,14 +45,12 @@ var _ webhook.Validator = &SeldonDeployment{}
 // ValidateCreate implements webhook.Validator so a webhook will be registered for the type
 func (r *SeldonDeployment) ValidateCreate() error {
 	seldondeploymentlog.Info("Validating v1alpha2 Webhook called for CREATE", "name", r.Name)
-
 	return r.Spec.ValidateSeldonDeployment()
 }
 
 // ValidateUpdate implements webhook.Validator so a webhook will be registered for the type
 func (r *SeldonDeployment) ValidateUpdate(old runtime.Object) error {
 	seldondeploymentlog.Info("Validating v1alpha2 webhook called for UPDATE", "name", r.Name)
-
 	return r.Spec.ValidateSeldonDeployment()
 }
