+++ operator/config/cert/kustomization.yaml
@@ -9,13 +9,6 @@ patchesJson6902:
 - target:
     group: admissionregistration.k8s.io
     version: v1beta1
-    kind: MutatingWebhookConfiguration
-    name: mutating-webhook-configuration
-  path: patch_webhook.yaml
-- target:
-    group: admissionregistration.k8s.io
-    version: v1beta1
     kind: ValidatingWebhookConfiguration
     name: validating-webhook-configuration
   path: patch_webhook.yaml
-
