+++ operator/config/manifests/bases/seldon-operator.clusterserviceversion.yaml
@@ -81,9 +81,6 @@ spec:
       - kind: HorizontalPodAutoscaler
         name: ""
         version: v2beta1
-      - kind: MutatingWebhookConfiguration
-        name: ""
-        version: v1beta1
       - kind: ValidatingWebhookConfiguration
         name: ""
         version: v1beta1
@@ -162,7 +159,6 @@ spec:
         - apiGroups:
           - admissionregistration.k8s.io
           resources:
-          - mutatingwebhookconfigurations
           - validatingwebhookconfigurations
           verbs:
           - get
@@ -386,7 +382,6 @@ spec:
         - apiGroups:
           - admissionregistration.k8s.io
           resources:
-          - mutatingwebhookconfigurations
           - validatingwebhookconfigurations
           verbs:
           - get
