+++ operator/config/namespaced/webhook-config.yaml
@@ -1,5 +1,3 @@
 varReference:
 - path: webhooks/namespaceSelector/matchLabels/seldon.io\/controller-id
-  kind: MutatingWebhookConfiguration
-- path: webhooks/namespaceSelector/matchLabels/seldon.io\/controller-id
   kind: ValidatingWebhookConfiguration
